circuit-macros: Refactor macro implementation to rely heavily on helpers
